refine test cases, parameters and data for flex type
