--- interactive_engine/compiler/src/main/resources/flex_bench/parameters
@@ -1,59 +1,0 @@
-test_case_name|parameters|results
-compare_int32_int32|12|12
-compare_uint32_uint32|+13|+13
-compare_uint32_int32|12|+12
-compare_uint32_int32_overflow|-12|empty
-compare_uint64_uint64|+14L|+14L
-compare_int64_int64|14L|14L
-compare_uint64_int64|14L|+14L
-compare_uint64_int64_overflow|-14L|empty
-compare_int32_int64|15L|15
-compare_int32_int64_overflow|2147483648L|empty
-compare_float_float|16.0f|16.0f
-compare_double_double|16.12d|16.12d
-compare_float_double|16.10d|16.1f
-compare_float_double_loss|16.12d|empty
-compare_int32_double|16.0d|16
-compare_char_text|'abcd'|'abcd'
-compare_char_long_text|'abcde'|'abcd'
-compare_varchar_long_text|'abcde'|'abcd'
-compare_string_long_text|'abcde...'|'abcde...'
-compare_date32_i32|1200|1200
-compare_timestamp_i64|1678322400000|1678322400000
-
-plus_int32_int32|-12,13|1
-plus_int32_int32_overflow|2147483647,1|overflow
-plus_int32_uint32|12,+13|+25
-plus_int32_uint32_overflow|-12,+13|overflow
-plus_int32_int64|12,14L|26L
-plus_int32_double|12,16.23d|28.23d
-plus_float_double|12.0f,16.23d|28.23d
-
-minus_int32_int32|12,13|-1
-minus_int32_int32_overflow|2147483647,-1|overflow
-minus_int32_uint32|13,+12|+1
-minus_int32_uint32_overflow|12,+13|overflow
-minus_int32_int64|12,13L|-1L
-minus_int32_double|12,13.12d|-1.12d
-minus_float_double|12.0f,13.12d|-1.12d
-
-multiply_int32_int32|1,2|2
-multiply_int32_int32_overflow|2147483647,2|overflow
-multiply_int32_uint32|1,+2|+2
-multiply_int32_uint32_overflow|-100,+100|overflow
-multiply_int32_int64|1,2L|2L
-multiply_int32_double|2,2.12d|4.24d
-multiply_float_double|2.0f,2.12d|4.24d
-
-divide_int32_int32|2,1|2
-divide_int32_int32_overflow|-2147483648,-1|overflow
-divide_int32_uint32|2,+1|+2
-divide_uint32_int32_overflow|+1000,-2147483648|overflow
-divide_int32_int64|2,1L|2L
-divide_int32_double|2,1.12d|1.77d
-divide_float_double|2.0f,1.12d|1.78d
-divide_float_double_NaN|2.0f,0.0d|NaN
-
-get_knows_between_two_persons|1,2|1,2
-
-
