+++ interactive_engine/compiler/src/test/resources/flex_bench/data/knows.csv
@@ -1,0 +1,4 @@
+source_person_id,destination_person_id
+32,33
+33,34
+32,34
