+++ interactive_engine/compiler/src/test/resources/flex_bench/parameters
@@ -1,0 +1,72 @@
+test_case_name|parameters|results
+compare_int32_int32|32|32
+compare_uint32_uint32|+12345|+12345
+compare_uint32_int32|12345|+12345
+// 4294967284
+compare_uint32_int32_overflow|-12|empty
+compare_uint64_uint64|+9876543210L|+9876543210L
+compare_int64_int64|1234567890L|1234567890L
+compare_uint64_int64|9876543210L|+9876543210L
+// 18446744073709551602
+compare_uint64_int64_overflow|-14L|empty
+compare_int32_int64|32L|32
+// -2147483648
+compare_int32_int64_overflow|2147483648L|empty
+compare_float_float|3.14f|3.14f
+compare_double_double|3.14159d|3.14159d
+compare_float_double|3.14d|3.14f
+compare_float_double_loss|3.14159d|empty
+compare_int32_double|32.0d|32
+compare_char_text|"abcd"|"abcd"
+compare_char_long_text|"abcde"|"abcd"
+compare_varchar_long_text|"abcde"|"abcd"
+compare_string_long_text|"Long text with no truncation"|"Long text with no truncation"
+// 2025-02-13
+compare_date32_i32|20132|20132
+// 2025-02-13T13:45:01
+compare_timestamp_i64|1739454301000L|1739454301000L
+
+plus_int32_int32|12,13|25
+plus_int32_int32_overflow|2147483647,1|overflow
+plus_int32_uint32_int32|12,+13|25
+plus_int32_uint32_uint32|10,+2147483647|+2147483657
+plus_int32_uint32_overflow|1,+4294967295|overflow
+plus_int32_int64|12,14L|26L
+plus_int32_int64_overflow|10,9223372036854775807L|overflow
+plus_int32_double|12,16.23d|28.23d
+plus_float_double|12.0f,16.23d|28.23d
+
+minus_int32_int32|12,13|-1
+minus_int32_int32_overflow|2147483647,-1|overflow
+minus_int32_uint32_int32|12,+13|-1
+minus_uint32_int32_uint32|+2147483647,-10|+2147483657
+minus_uint32_int32_overflow|+4294967295,-1|overflow
+minus_int32_int64|12,13L|-1L
+minus_int64_int32_overflow|9223372036854775807L,-10|overflow
+minus_int32_double|12,13.12d|-1.12d
+minus_float_double|12.0f,13.12d|-1.12d
+
+multiply_int32_int32|1,2|2
+multiply_int32_int32_overflow|2147483647,2|overflow
+multiply_int32_uint32_int32|1,+2|2
+multiply_int32_uint32_uint32|4,+1000000000|+4000000000
+multiply_int32_uint32_overflow|4,+2000000000|overflow
+multiply_int32_int64|1,2L|2L
+multiply_int32_int64_overflow|2,9223372036854775807L|overflow
+multiply_int32_double|2,2.12d|4.24d
+multiply_float_double|2.0f,2.12d|4.24d
+
+divide_int32_int32|2,1|2
+divide_int32_int32_overflow|-2147483648,-1|overflow
+divide_int32_uint32_int32|2,+1|2
+divide_uint32_int32_uint32|+4294967295,1|+4294967295
+divide_uint32_int32_overflow|+4294967295,-1|overflow
+divide_int32_int64|2,1L|2L
+divide_int64_int32_overflow|-9223372036854775808L,-1|overflow
+divide_int32_double|2,1.12d|1.77d
+divide_float_double|2.0f,1.12d|1.78d
+divide_float_double_NaN|2.0f,0.0d|NaN
+
+get_knows_between_two_persons|32,33|32,33
+
+
